+++ lib/ts/recipe/session/api/implementation.ts
@@ -4,6 +4,7 @@ import NormalisedURLPath from "../../../normalisedURLPath";
 import { SessionContainerInterface } from "../types";
 import { GeneralErrorResponse } from "../../../types";
 import { getRequiredClaimValidators } from "../utils";
+import { clearSessionFromCookie } from "../cookieAndHeaders";
 
 export default function getAPIInterface(): APIInterface {
     return {
@@ -67,6 +68,7 @@ export default function getAPIInterface(): APIInterface {
         },
 
         signOutPOST: async function ({
+            options,
             session,
             userContext,
         }: {
@@ -82,6 +84,9 @@ export default function getAPIInterface(): APIInterface {
             if (session !== undefined) {
                 await session.revokeSession(userContext);
             }
+
+            clearSessionFromCookie(options.config, options.res);
+
             return {
                 status: "OK",
             };